# Add eBPF module to FIM whodata