--- a/packages/debs/SPECS/wazuh-agent/debian/wazuh-agent.lintian-overrides
+++ b/packages/debs/SPECS/wazuh-agent/debian/wazuh-agent.lintian-overrides
@@ -7,4 +7,4 @@ wazuh-agent: non-standard-dir-in-var
 wazuh-agent: file-in-unusual-dir
 wazuh-agent: hardening-no-fortify-functions
 wazuh-agent: hardening-no-relro
-wazuh-agent: binary-from-other-architecture var/ossec/bin/modern.bpf.o
+wazuh-agent: binary-from-other-architecture var/ossec/lib/modern.bpf.o