--- a/src/Makefile
+++ b/src/Makefile
@@ -85,6 +85,7 @@ EXTERNAL_JEMALLOC=external/jemalloc/
 ifeq (${uname_S},Linux)
 EXTERNAL_RPM=external/rpm/
 EXTERNAL_POPT=external/popt/
+EXTERNAL_LIBBPF=external/libbpf-bootstrap/
 endif
 # XXX Becareful NO EXTRA Spaces here
 PG_CONFIG?=pg_config
@@ -949,6 +950,7 @@ ROCKSDB_LIB = $(EXTERNAL_ROCKSDB)build/librocksdb.so
 FLATBUFFERS_LIB = $(EXTERNAL_FLATBUFFERS)build/libflatbuffers.a
 LZMA_LIB = $(EXTERNAL_LZMA)build/liblzma.a
 LUA_LIB = ${EXTERNAL_LUA}install/lib/liblua.a
+LIBBPF_LIB = $(EXTERNAL_LIBBPF)build/modern.bpf.o
 
 EXTERNAL_LIBS := $(JSON_LIB) $(ZLIB_LIB) $(OPENSSL_LIB) $(CRYPTO_LIB) $(SQLITE_LIB) $(LIBYAML_LIB) $(LIBPCRE2_LIB)
 
@@ -1001,6 +1003,12 @@ endif
 
 ifeq (${TARGET},server)
 external: $(ROCKSDB_LIB)
+endif
+
+ifeq (${uname_S},Linux)
+ifneq ($(CHECK_CENTOS5),YES)
+external: ${LIBBPF_LIB}
+endif
 endif
 
 ifneq (${TARGET},winagent)
@@ -1059,6 +1067,11 @@ endif
 endif
 endif
 endif
+
+#### libbpf-bootstrap #########
+
+$(LIBBPF_LIB):
+	cd $(EXTERNAL_LIBBPF) && cmake -B build && cmake --build build
 
 ### rocks db ###
 
@@ -1369,8 +1382,8 @@ TAR := tar -xf
 GUNZIP := gunzip
 GZIP := gzip
 CURL := curl -so
-DEPS_VERSION = 40
-RESOURCES_URL_BASE := https://packages.wazuh.com/deps/
+DEPS_VERSION = 27598
+RESOURCES_URL_BASE := https://packages-dev.wazuh.com/deps/
 RESOURCES_URL := $(RESOURCES_URL_BASE)$(DEPS_VERSION)
 CPYTHON := cpython
 PYTHON_SOURCE := no
@@ -1449,17 +1462,11 @@ PRECOMPILED_RES := $(PRECOMPILED_OS)$(PRECOMPILED_ARCH)
 endif
 
 # Agent dependencies
-EXTERNAL_RES := cJSON curl libdb libffi libyaml openssl procps sqlite zlib audit-userspace msgpack bzip2 nlohmann googletest libpcre2 libplist pacman libarchive popt lua rpm rocksdb lzma cpp-httplib benchmark
+EXTERNAL_RES := cJSON curl libdb libffi libyaml openssl procps sqlite zlib audit-userspace msgpack bzip2 nlohmann googletest libpcre2 libplist pacman libarchive popt lua rpm rocksdb lzma cpp-httplib benchmark libbpf-bootstrap
 ifneq (${TARGET},agent)
 ifneq (${TARGET},winagent)
 	# Manager extra dependency
 	EXTERNAL_RES += $(CPYTHON) jemalloc flatbuffers
-endif
-endif
-
-ifneq (${TARGET},server)
-ifeq (${TARGET},agent)
-       EXTERNAL_RES += libbpf-bootstrap
 endif
 endif
 
@@ -2740,6 +2747,7 @@ ifneq ($(wildcard external/*/*),)
 	rm -rf $(EXTERNAL_GOOGLE_TEST)build
 	rm -rf $(EXTERNAL_GOOGLE_BENCHMARK)build
 	rm -rf $(EXTERNAL_ROCKSDB)build
+	rm -rf $(EXTERNAL_LIBBPF)build
 	rm -rf $(EXTERNAL_FLATBUFFERS)build
 	rm -rf $(EXTERNAL_LZMA)build
 	-cd ${EXTERNAL_LIBPLIST} && ${MAKE} clean && rm -rf bin/*