--- a/src/Makefile
+++ b/src/Makefile
@@ -1071,7 +1071,7 @@ endif
 #### libbpf-bootstrap #########
 
 $(LIBBPF_LIB):
-	cd $(EXTERNAL_LIBBPF) && cmake -B build && cmake --build build
+	cd $(EXTERNAL_LIBBPF) && mkdir -p build && cd build && cmake .. && ${MAKE}
 
 ### rocks db ###