--- a/src/error_messages/error_messages.h
+++ b/src/error_messages/error_messages.h
@@ -476,6 +476,9 @@
 #define FIM_AUDITPOL_ATTEMPT_FAIL                   "(6955): Auditpol command failed, attempt number: %d"
 #define FIM_AUDITPOL_FINAL_FAIL                     "(6956): After %d attempts the Auditpol command could not be executed successfully."
 
+#define FIM_ERROR_EBPF_NOT_SUPPORTED                "(6966): Wrong configuration. eBPF is not available on non-Linux systems."
+#define FIM_ERROR_EBPF_HEALTHCHECK                  "(6967): The eBPF healthcheck has failed. Switching all whodata eBPF configuration to audit."
+
 /* Wazuh Logtest error messsages */
 #define LOGTEST_ERROR_BIND_SOCK                     "(7300): Unable to bind to socket '%s'. Errno: (%d) %s"
 #define LOGTEST_ERROR_ACCEPT_CONN                   "(7301): Failure to accept connection. Errno: %s"