--- a/src/error_messages/information_messages.h
+++ b/src/error_messages/information_messages.h
@@ -63,6 +63,8 @@
 #define FIM_AUDIT_CREATED_RULE_FILE         "(6045): Created audit rules file, due to audit immutable mode rules will be loaded in the next reboot."
 #define FIM_AUDIT_QUEUE_SIZE                "(6046): Internal audit queue size set to '%d'."
 
+#define FIM_EBPF_INIT                       "(6047): Initializing eBPF driver for FIM whodata."
+
 /* wazuh-logtest information messages */
 #define LOGTEST_INITIALIZED                 "(7200): Logtest started"
 #define LOGTEST_DISABLED                    "(7201): Logtest disabled"