--- a/src/init/inst-functions.sh
+++ b/src/init/inst-functions.sh
@@ -897,25 +897,25 @@ InstallCommon()
        		if ([ "X${DIST_NAME}" = "Xrhel" ] || [ "X${DIST_NAME}" = "Xcentos" ] || [ "X${DIST_NAME}" = "XCentOS" ]) && [ ${DIST_VER} -le 5 ]; then
        		    chcon -t textrel_shlib_t ${INSTALLDIR}/lib/libfimebpf.so
        		fi
-	fi
+      fi
 
-	if [ -f external/libbpf_boostrap/build/libbpf/libbpf.so ]
-    	then
-       		${INSTALL} -m 0750 -o root -g ${WAZUH_GROUP} external/libbpf_boostrap/build/libbpf/libbpf.so ${INSTALLDIR}/lib
+      if [ -f external/libbpf-bootstrap/build/libbpf/libbpf.so ]
+          then
+              ${INSTALL} -m 0750 -o root -g ${WAZUH_GROUP} external/libbpf-bootstrap/build/libbpf/libbpf.so ${INSTALLDIR}/lib
 
-       		if ([ "X${DIST_NAME}" = "Xrhel" ] || [ "X${DIST_NAME}" = "Xcentos" ] || [ "X${DIST_NAME}" = "XCentOS" ]) && [ ${DIST_VER} -le 5 ]; then
-       		    chcon -t textrel_shlib_t ${INSTALLDIR}/lib/libbpf.so
-       		fi
-	fi
+              if ([ "X${DIST_NAME}" = "Xrhel" ] || [ "X${DIST_NAME}" = "Xcentos" ] || [ "X${DIST_NAME}" = "XCentOS" ]) && [ ${DIST_VER} -le 5 ]; then
+                  chcon -t textrel_shlib_t ${INSTALLDIR}/lib/libbpf.so
+              fi
+      fi
 
-	if [ -f external/libbpf_boostrap/build/modern.bpf.o ]
-    	then
-       		${INSTALL} -m 0750 -o root -g ${WAZUH_GROUP} external/libbpf_boostrap/build/modern.bpf.o ${INSTALLDIR}/bin
+      if [ -f external/libbpf-bootstrap/build/modern.bpf.o ]
+          then
+              ${INSTALL} -m 0750 -o root -g ${WAZUH_GROUP} external/libbpf-bootstrap/build/modern.bpf.o ${INSTALLDIR}/bin
 
-       		if ([ "X${DIST_NAME}" = "Xrhel" ] || [ "X${DIST_NAME}" = "Xcentos" ] || [ "X${DIST_NAME}" = "XCentOS" ]) && [ ${DIST_VER} -le 5 ]; then
-       		    chcon -t textrel_shlib_t ${INSTALLDIR}/bin/modern.bpf.o
-       		fi
-	fi
+              if ([ "X${DIST_NAME}" = "Xrhel" ] || [ "X${DIST_NAME}" = "Xcentos" ] || [ "X${DIST_NAME}" = "XCentOS" ]) && [ ${DIST_VER} -le 5 ]; then
+                  chcon -t textrel_shlib_t ${INSTALLDIR}/bin/modern.bpf.o
+              fi
+      fi
     fi
 
     if [ ${NUNAME} = 'Darwin' ]