--- a/src/syscheckd/src/ebpf/tests/CMakeLists.txt
+++ b/src/syscheckd/src/ebpf/tests/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(unit)