… .github/actions/check_files/deb_linux_agent_amd64.csv
… .github/actions/check_files/deb_linux_agent_i386.csv
… .github/actions/check_files/manager_base.csv
… .github/actions/check_files/rpm_linux_agent_amd64.csv
… .github/actions/check_files/rpm_linux_agent_i386.csv
… src/syscheckd/CMakeLists.txt
… src/syscheckd/include/syscheck.h
… src/syscheckd/src/config.c
… src/syscheckd/src/create_db.c
… src/syscheckd/src/ebpf/CMakeLists.txt
… src/syscheckd/src/ebpf/include/bounded_queue.hpp
… src/syscheckd/src/ebpf/include/bpf_helpers.h
… src/syscheckd/src/ebpf/include/ebpf_whodata.h
… src/syscheckd/src/ebpf/include/ebpf_whodata.hpp
… src/syscheckd/src/ebpf/include/vmlinux.h
… src/syscheckd/src/ebpf/include/wrapper_bpf.h
… src/syscheckd/src/ebpf/src/ebpf_whodata.cpp
… src/syscheckd/src/ebpf/src/modern.bpf.c
… src/syscheckd/src/ebpf/tests/unit/CMakeLists.txt
… src/syscheckd/src/ebpf/tests/unit/bounded_queue_test.cpp
… src/syscheckd/src/main.c
… src/syscheckd/src/run_check.c
… src/syscheckd/src/run_realtime.c
… src/syscheckd/src/syscheck.c
… src/syscheckd/src/whodata/audit_rule_handling.c
… src/syscheckd/src/whodata/syscheck_audit.c
… src/syscheckd/src/whodata/win_whodata.c
… src/unit_tests/agent.cmake
… src/unit_tests/config_files/test_syscheck.conf
… src/unit_tests/config_files/test_syscheck2.conf