|Related issue|
|---|
|https://github.com/wazuh/wazuh/issues/27879|

## Description

Hi team, 

This development enables a new FIM who-data mode based on eBPF program.

eBPF programs are a piece of code that runs in the kernel and we use it to extract events straight from it, allowing us to not depend on an external dependency as `auditd` and letting us extract the generated events in a faster way.

The events that will be monitored are:
* `vfs_open`: this system call lets us know when a new file has been created.
* `security_inode_setattr`: this one lets us detect when a file has been modified.
* `vfs_unlink`: detects when a file has been removed from the letssystem.

This development uses a kernel structure called `ring_buffer`, which is the data structure we use to pass events from the kernel to the user space, where our who-data engine analyzes them.

This data structure is present in kernels starting from `5.8` version, so this development will be only leveraged if the system has a `+5.8` kernel version. If the kernel version is older, but this who-data mode is configured, the `healthcheck` will detect it, letting the user know the reason why this mode is not available and falling back to `audit` mode.

**Considerations**
Since this who-data mode retrieves the events way faster than `audit`, memory consumption can show some momentaneous peaks. This behavior is most likely to be seen in machines with 1 and 2 CPUs which will be harder for them to consume all the fetched events from the kernel.

## Configuration options

To enable this who-data mode, the following configuration block has been added:

```xml
<syscheck>
  <directories whodata="yes">/home/user/documents</directories>

  <whodata>
    <provider>ebpf</provider>
  </whodata>
</syscheck>
```

Inside the `<syscheck>` configuration block and the `<whodata>` tag, the new field `<provider>` is added, indicating the selected who-data mode. It's worth mentioning, both who-data modes `audit` which is the default one, and the newly added `ebpf` can not coexist, so every who-data monitored folder will be monitored using the same method.

If whodata is enabled but no <provider> is specified, `audit` will be set by default.

It's also important to mention the following configuration block:

```xml
<whodata>
    <queue_size>50000</queue_size>
</whodata>
```

Due to the high speed, the `ebpf` detected events are fetched, it's important to fine-tune the `queue_size` field to give it enough size to handle a big burst of events that may be detected from the kernel in a, for instance, massive deletion event.

A `queue_size` value increase will prevent events from being lost when a large amount of events is detected, but it may also carry an increase in the memory consumption metric of the agent if the compute capabilities of the monitored endpoint are limited.

## Logs/Alerts example

The alerts structure is kept with no changes; the same data fields extracted using `audit` are being extracted using `eBPF`.

<details><summary>Detected alerts</summary>

```
** Alert 1743067416.30801: - ossec,syscheck,syscheck_entry_added,syscheck_file,pci_dss_11.5,gpg13_4.11,gdpr_II_5.1.f,hipaa_164.312.c.1,hipaa_164.312.c.2,nist_800_53_SI.7,tsc_PI1.4,tsc_PI1.5,tsc_CC6.1,tsc_CC6.8,tsc_CC7.2,tsc_CC7.3,
2025 Mar 27 09:23:36 (agent1-ubu22) any->syscheck
Rule: 554 (level 5) -> 'File added to the system.'
File '/test_ebpf/tes' added
Mode: whodata

Attributes:
 - Size: 0
 - Permissions: rw-r--r--
 - Date: Thu Mar 27 09:23:36 2025
 - Inode: 2359298
 - User: root (0)
 - Group: root (0)
 - MD5: d41d8cd98f00b204e9800998ecf8427e
 - SHA1: da39a3ee5e6b4b0d3255bfef95601890afd80709
 - SHA256: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
 - (Audit) User name: root
 - (Audit) Group name: root
 - (Audit) Process id: 2277
 - (Audit) Process name: touch
 - (Audit) Process cwd: /home/vagrant
 - (Audit) Parent process name: bash
 - (Audit) Parent process id: 1373
 - (Audit) Parent process cwd: /home/vagrant

** Alert 1743067428.31751: - ossec,syscheck,syscheck_entry_modified,syscheck_file,pci_dss_11.5,gpg13_4.11,gdpr_II_5.1.f,hipaa_164.312.c.1,hipaa_164.312.c.2,nist_800_53_SI.7,tsc_PI1.4,tsc_PI1.5,tsc_CC6.1,tsc_CC6.8,tsc_CC7.2,tsc_CC7.3,
2025 Mar 27 09:23:48 (agent1-ubu22) any->syscheck
Rule: 550 (level 7) -> 'Integrity checksum changed.'
File '/test_ebpf/tes' modified
Mode: whodata
Changed attributes: size,mtime,md5,sha1,sha256
Size changed from '0' to '6'
Old modification time was: '1743067416', now it is '1743067428'
Old md5sum was: 'd41d8cd98f00b204e9800998ecf8427e'
New md5sum is : 'b1946ac92492d2347c6235b4d2611184'
Old sha1sum was: 'da39a3ee5e6b4b0d3255bfef95601890afd80709'
New sha1sum is : 'f572d396fae9206628714fb2ce00f72e94f2258f'
Old sha256sum was: 'e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855'
New sha256sum is : '5891b5b522d5df086d0ff0b110fbd9d21bb4fc7163af34d08286a2e846f6be03'

Attributes:
 - Size: 6
 - Permissions: rw-r--r--
 - Date: Thu Mar 27 09:23:48 2025
 - Inode: 2359298
 - User: root (0)
 - Group: root (0)
 - MD5: b1946ac92492d2347c6235b4d2611184
 - SHA1: f572d396fae9206628714fb2ce00f72e94f2258f
 - SHA256: 5891b5b522d5df086d0ff0b110fbd9d21bb4fc7163af34d08286a2e846f6be03
 - (Audit) User name: root
 - (Audit) Group name: root
 - (Audit) Process id: 1373
 - (Audit) Process name: bash
 - (Audit) Process cwd: /home/vagrant
 - (Audit) Parent process name: su
 - (Audit) Parent process id: 1372

** Alert 1743067433.33195: - ossec,syscheck,syscheck_entry_deleted,syscheck_file,pci_dss_11.5,gpg13_4.11,gdpr_II_5.1.f,hipaa_164.312.c.1,hipaa_164.312.c.2,nist_800_53_SI.7,tsc_PI1.4,tsc_PI1.5,tsc_CC6.1,tsc_CC6.8,tsc_CC7.2,tsc_CC7.3,
2025 Mar 27 09:23:53 (agent1-ubu22) any->syscheck
Rule: 553 (level 7) -> 'File deleted.'
File '/test_ebpf/tes' deleted
Mode: whodata

Attributes:
 - Size: 6
 - Permissions: rw-r--r--
 - Date: Thu Mar 27 09:23:48 2025
 - Inode: 2359298
 - User: root (0)
 - Group: root (0)
 - MD5: b1946ac92492d2347c6235b4d2611184
 - SHA1: f572d396fae9206628714fb2ce00f72e94f2258f
 - SHA256: 5891b5b522d5df086d0ff0b110fbd9d21bb4fc7163af34d08286a2e846f6be03
 - (Audit) User name: root
 - (Audit) Group name: root
 - (Audit) Process id: 2278
 - (Audit) Process name: rm
 - (Audit) Process cwd: /home/vagrant
 - (Audit) Parent process name: bash
 - (Audit) Parent process id: 1373
 - (Audit) Parent process cwd: /home/vagrant
```

</details>

## Tests

To ensure this development works as expected and runs on the supported kernel versions, the following issues have been worked on to extensively test this new feature.

* https://github.com/wazuh/wazuh/issues/28789: this issue has tested that the development works in kernels +5.8 by default but also in older kernels that have been upgraded to a +5.8 version to satisfy the version restriction.

* https://github.com/wazuh/wazuh/issues/28811; this issue studies a performance comparison between `audit` and `ebpf` who-data modes, showcasing the compute duration during a high-stress situation is lower using `ebpf`, meaning the FIM events are detected much faster using this new mode.

* https://github.com/wazuh/wazuh/issues/28631: this issue shows the performance compared to other open source projects that also make use of eBPF.

* https://github.com/wazuh/internal-devel-requests/issues/2178: this issue shows several footprint tests launched to visualize the consumption of our eBPF development.

- [x] Coverity
- [x] Valgrind
- [x] scan-build
- [x] Linux packages generated
  - [x] `deb`: amd64, arm64, i386, arm32 and PPC64. 
  - [x] `rpm`: amd64, arm64, i386, arm32 and PPC64.