@jotacarma90 2025-03-27
# Full code revision 🟢 
LGTM

# Final testing in Ubuntu 22 🟢 
```
2025/03/27 14:06:17 wazuh-syscheckd[65365] main.c:210 at main(): INFO: (6003): Monitoring path: '/test', with options 'size | permissions | owner | group | mtime | inode | hash_md5 | hash_sha1 | hash_sha256 | whodata'.

2025/03/27 14:06:17 wazuh-syscheckd[65365] syscheck.c:320 at check_ebpf_availability(): INFO: (6047): Initializing eBPF driver for FIM whodata.
2025/03/27 14:06:17 wazuh-syscheckd[65365] logging_helper.c:65 at loggingFunction(): DEBUG: (6376): All functions loaded successfully from libbpf.so.

2025/03/27 14:06:20 wazuh-syscheckd[65365] logging_helper.c:56 at loggingFunction(): INFO: (6048): Healthcheck for eBPF FIM whodata module success.

2025/03/27 14:07:08 wazuh-syscheckd[65365] run_check.c:125 at send_syscheck_msg(): DEBUG: (6321): Sending FIM event: {"data":{"attributes":{"checksum":"6b9c9087849d1afdeefd0a7754a7337d64a8b091","gid":"0","group_name":"root","hash_md5":"d41d8cd98f00b204e9800998ecf8427e","hash_sha1":"da39a3ee5e6b4b0d3255bfef95601890afd80709","hash_sha256":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","inode":655364,"mtime":1743084428,"perm":"rw-r--r--","size":0,"type":"file","uid":"0","user_name":"root"},"mode":"whodata","path":"/test/file1","timestamp":1743084428,"type":"added","version":"2.0","audit":{"user_id":"0","user_name":"root","process_name":"touch","process_id":66687,"cwd":"/home/vagrant/wazuh/src","group_id":"0","group_name":"root","parent_name":"bash","parent_cwd":"/home/vagrant/wazuh/src","ppid":3338}},"type":"event"}
2025/03/27 14:07:12 wazuh-syscheckd[65365] run_check.c:125 at send_syscheck_msg(): DEBUG: (6321): Sending FIM event: {"data":{"attributes":{"checksum":"769b48b672a3eab204ba9b88244b5fd98e52878d","gid":"0","group_name":"root","hash_md5":"d41d8cd98f00b204e9800998ecf8427e","hash_sha1":"da39a3ee5e6b4b0d3255bfef95601890afd80709","hash_sha256":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","inode":655364,"mtime":1743084432,"perm":"rw-r--r--","size":0,"type":"file","uid":"0","user_name":"root"},"changed_attributes":["mtime"],"mode":"whodata","old_attributes":{"checksum":"6b9c9087849d1afdeefd0a7754a7337d64a8b091","gid":"0","group_name":"root","hash_md5":"d41d8cd98f00b204e9800998ecf8427e","hash_sha1":"da39a3ee5e6b4b0d3255bfef95601890afd80709","hash_sha256":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","inode":655364,"mtime":1743084428,"perm":"rw-r--r--","size":0,"type":"file","uid":"0","user_name":"root"},"path":"/test/file1","timestamp":1743084432,"type":"modified","version":"2.0","audit":{"user_id":"0","user_name":"root","process_name":"touch","process_id":66688,"cwd":"/home/vagrant/wazuh/src","group_id":"0","group_name":"root","parent_name":"bash","parent_cwd":"/home/vagrant/wazuh/src","ppid":3338}},"type":"event"}
2025/03/27 14:07:15 wazuh-syscheckd[65365] run_check.c:125 at send_syscheck_msg(): DEBUG: (6321): Sending FIM event: {"type":"event","data":{"path":"/test/file1","version":2,"mode":"whodata","type":"deleted","timestamp":1743084432,"attributes":{"type":"file","size":0,"perm":"rw-r--r--","uid":"0","gid":"0","user_name":"root","group_name":"root","inode":655364,"mtime":1743084432,"hash_md5":"d41d8cd98f00b204e9800998ecf8427e","hash_sha1":"da39a3ee5e6b4b0d3255bfef95601890afd80709","hash_sha256":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","checksum":"769b48b672a3eab204ba9b88244b5fd98e52878d"},"audit":{"user_id":"0","user_name":"root","process_name":"rm","process_id":66689,"cwd":"/home/vagrant/wazuh/src","group_id":"0","group_name":"root","parent_name":"bash","parent_cwd":"/home/vagrant/wazuh/src","ppid":3338}}}
```